Security group report
Prints out the number of security group rules unique by ip_range and the number of security groups with an allow any rule.
Useful if you need to report on security in a context more suited to old school firewalls. Could be extended to a proper report with much more information (might do this later).